+++ frontend/src/entities/consent/index.js
@@ -1,0 +1,4 @@
+export { default as ConsentProvider } from './model/ConsentProvider'
+export { useConsent } from './model/consent-context'
+export { useConsentedColorScheme } from './model/use-consented-color-scheme'
+export { EMPTY_CONSENT, hasPreferenceConsent } from './lib/consent-storage'